Added modal bottom sheet that is triggered by tab buttons and user pic.
Full implementation of specific tab actions to be done still.
