+++ lib/official_themes.dart
@@ -1,4 +1,8 @@
 import 'package:flutter/material.dart';
 
-const Color upvoteColor = Color.fromARGB(255, 255, 69, 0);
-const Color downvoteColor = Color.fromARGB(255, 113, 147, 255);
+class LeafColors {
+  LeafColors._();
+
+  static const Color upvote_color = const Color.fromARGB(255, 255, 69, 0);
+  static const Color downvote_color = const Color.fromARGB(255, 113, 147, 255);
+}
