+++ lib/user_services.dart
@@ -1,0 +1,14 @@
+/*
+ * Interface currently without implementation.
+ * Objective is to provide user profile information for widgets
+ */
+
+abstract class UserInformationManager {
+  String imageUrl;
+  String username;
+
+  factory UserInformationManager() {
+    // TODO provide default implementation
+    return null;
+  }
+}
